Added the "Smart"-AI
Added a "Smart"-AI which bases its moves based on a formula, which calculates based on the individual's genotype, whether a certain move should be beneficial or not. Is able to greatly outperform "Random" or shoot itself in the foot based on the genotype.
The next step is to perform either gradient descent, or some form of EA which will make sure the model learns.