Coupled inherritance has been removed
Coupled inherritance was removed since the hypothesis followed that that was where the chaos from the graphs came from, however, this is now known not to be the case. On top of this, a new evident problem has popped up, which makes it difficult to see the natural selection / evolution through the species, due to the wins not steadily increasing.
The next step is to make it so that numerous models are trained at the same time against random computers, and those that perform best are able to reproduce into the next generation. So that the win-rate and thus the performance shows steady growth over time.