Resolved some destructive bugs
Resolved the playGame()-method returning the wrong players' win due to an error in the length of the lists.
Resolved the calculateScore()-method returning the wrong scores for boards.
Resolved the main()-function, where the wrong information was called, which resulted in one model completing all of the games. Creating quite the barf for data out of each of the iterations.